refactor(config): extract proxy process-state ownership
